+++ tests/codegen/builtin_ncpus.cpp
@@ -1,0 +1,14 @@
+#include "common.h"
+
+namespace bpftrace {
+namespace test {
+namespace codegen {
+
+TEST(codegen, builtin_ncpus)
+{
+  test("BEGIN { @x = ncpus; exit(); }", NAME);
+}
+
+} // namespace codegen
+} // namespace test
+} // namespace bpftrace